Software/ArtelSimulations/requirements.txt
PyAutoGUI==0.9.50
numpy==1.18.3
matplotlib==3.2.2
scipy==1.5.0
pygcode==0.2.1
pyaudio==0.2.11
pyserial==3.5
